refactor: customElementsと自作コンポーネントを区別しない
