+++ src/LinkInternal.ts
@@ -1,4 +1,4 @@
-import { toState } from '../state';
+import { toState } from './state';
 
 export function defineLinkInternal(invokeRoute: () => void): void {
   class LinkInternal extends HTMLElement {
